refactored UI into its own class
