+++ .gitignore
@@ -2,3 +2,4 @@ src/conf/__pycache__/
 src/board/__pycache__/
 src/__pycache__/
 ui/MacroKeys_Configurator.glade~
+src/configurator/__pycache__/
